requirements.txt
aiomysql==0.1.1
alembic==1.10.3
anyio==3.6.2
asyncio==3.4.3
black==23.3.0
click==8.1.3
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
pydantic==1.10.7
PyMySQL==1.0.3
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==2.0.9
SQLAlchemy-Utils==0.41.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
